requirements.txt
authlib
python-dateutil
redis
requests
pylint
slackeventsapi
sqlalchemy
marshmallow-sqlalchemy>=0.21.0
marshmallow>=3.3.0
flask==1.0.2
flask_sqlalchemy==2.3.2
psycopg2-binary>=2.7.0
flask-migrate==2.4.0
flask-marshmallow>=0.10.1
